fix: Fixed minor flag bugs, moved to separate file
